refactor: Simplify file path and names
