+++ backend/src/fiches/controllers/fiches-action.controller.ts
@@ -6,7 +6,7 @@ import type { SupabaseJwtPayload } from '../../auth/models/supabase-jwt.models';
 import { getFichesActionSyntheseSchema } from '../models/get-fiches-action-synthese.response';
 import { getFichesActionFilterRequestSchema } from '../models/get-fiches-actions-filter.request';
 import { updateFicheActionRequestSchema } from '../models/update-fiche-action.request';
-import FichesActionSyntheseService from '../services/fiches-action-synthese.service';
+import SyntheseService from '../services/synthese.service';
 import FichesActionUpdateService from '../services/fiches-action-update.service';
 
 /**
@@ -27,7 +27,7 @@ export class UpdateFicheActionRequestClass extends createZodDto(
 @Controller('collectivites/:collectivite_id/fiches-action')
 export class FichesActionController {
   constructor(
-    private readonly fichesActionSyntheseService: FichesActionSyntheseService,
+    private readonly syntheseService: SyntheseService,
     private readonly fichesActionUpdateService: FichesActionUpdateService
   ) {}
 
@@ -42,7 +42,7 @@ export class FichesActionController {
     @Query() request: GetFichesActionFilterRequestClass,
     @TokenInfo() tokenInfo: SupabaseJwtPayload
   ) {
-    return this.fichesActionSyntheseService.getFichesActionSynthese(
+    return this.syntheseService.getFichesActionSynthese(
       collectiviteId,
       request,
       tokenInfo
@@ -59,7 +59,7 @@ export class FichesActionController {
     @Query() request: GetFichesActionFilterRequestClass,
     @TokenInfo() tokenInfo: SupabaseJwtPayload
   ) {
-    return this.fichesActionSyntheseService.getFichesAction(
+    return this.syntheseService.getFichesAction(
       collectiviteId,
       request,
       tokenInfo
