+++ backend/src/trpc/trpc.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { TrpcService } from './services/trpc.service';
+import { TrpcService } from './trpc.service';
 
 @Module({
   imports: [],
